Add tests for BeamChat
Move Beam tests into separate files in beam/ directory.
